Extract permissions dialog to its own component
