+++ i18n/es.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: i18next-conv\n"
-"POT-Creation-Date: 2020-06-04T11:36:59.775Z\n"
+"POT-Creation-Date: 2020-06-05T05:33:56.782Z\n"
 "PO-Revision-Date: 2018-10-25T09:02:35.143Z\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -32,6 +32,12 @@ msgstr "Cerrar"
 msgid "Program"
 msgstr "Programa"
 
+msgid "Access to Template Generation"
+msgstr "Acceso a la sección de generación de plantillas"
+
+msgid "Access to Settings and Themes"
+msgstr "Acceso a la configuración y los temas"
+
 msgid "Select Organisation Units on template generation"
 msgstr "Mostrar unidades organizativas en la sección de generación"
 
@@ -54,12 +60,6 @@ msgstr "Meses"
 
 msgid "Years"
 msgstr "Años"
-
-msgid "Access to Template Generation"
-msgstr "Acceso a la sección de generación de plantillas"
-
-msgid "Access to Settings and Themes"
-msgstr "Acceso a la configuración y los temas"
 
 msgid "Accessible to {{users}} users and {{userGroups}} user groups"
 msgstr "Accesible a {{users}} usuarios y {{userGroups}} grupos de usuarios"
