+++ i18n/fr.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: i18next-conv\n"
-"POT-Creation-Date: 2020-06-04T11:36:59.775Z\n"
+"POT-Creation-Date: 2020-06-05T05:33:56.782Z\n"
 "PO-Revision-Date: 2019-10-10T10:41:21.864Z\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -32,6 +32,12 @@ msgstr ""
 msgid "Program"
 msgstr ""
 
+msgid "Access to Template Generation"
+msgstr ""
+
+msgid "Access to Settings and Themes"
+msgstr ""
+
 msgid "Select Organisation Units on template generation"
 msgstr ""
 
@@ -51,12 +57,6 @@ msgid "Months"
 msgstr ""
 
 msgid "Years"
-msgstr ""
-
-msgid "Access to Template Generation"
-msgstr ""
-
-msgid "Access to Settings and Themes"
 msgstr ""
 
 msgid "Accessible to {{users}} users and {{userGroups}} user groups"
