+++ src/webapp/components/settings/SettingsFields.tsx
@@ -9,8 +9,6 @@ import {
     makeStyles,
     TextField,
 } from "@material-ui/core";
-import { D2Api } from "d2-api";
-import { ConfirmationDialog, ShareUpdate, Sharing, SharingRule } from "d2-ui-components";
 import React, { ChangeEvent, useCallback, useMemo, useState } from "react";
 import {
     DuplicateToleranceUnit,
@@ -18,22 +16,20 @@ import {
     OrgUnitSelectionSetting,
 } from "../../../domain/entities/AppSettings";
 import i18n from "../../../locales";
-import { useAppContext } from "../../contexts/api-context";
-import Settings, { PermissionSetting, PermissionType } from "../../logic/settings";
+import Settings, { PermissionSetting } from "../../logic/settings";
 import { Select, SelectOption } from "../select/Select";
 import DataElementsFilterDialog from "./DataElementsFilterDialog";
+import PermissionsDialog from "./PermissionsDialog";
 
 export interface SettingsFieldsProps {
     settings: Settings;
     onChange: (settings: Settings) => Promise<void>;
 }
 
-export default function SettingsFields(props: SettingsFieldsProps) {
-    const { settings, onChange } = props;
+export default function SettingsFields({ settings, onChange }: SettingsFieldsProps) {
     const classes = useStyles();
-    const { api } = useAppContext();
 
-    const [sharingDialogType, setSharingDialogType] = useState<PermissionSetting | null>(null);
+    const [permissionsType, setPermissionsType] = useState<PermissionSetting | null>(null);
     const [isExclusionDialogVisible, showExclusionDialog] = useState<boolean>(false);
 
     const setModel = useCallback(
@@ -101,53 +97,6 @@ export default function SettingsFields(props: SettingsFieldsProps) {
         []
     );
 
-    const search = React.useCallback((query: string) => searchUsers(api, query), [api]);
-
-    const buildMetaObject = useCallback(
-        (setting: PermissionSetting) => {
-            const displayName =
-                setting === "generation"
-                    ? i18n.t("Access to Template Generation")
-                    : i18n.t("Access to Settings and Themes");
-
-            const buildSharings = (type: PermissionType) =>
-                settings.getPermissions(setting, type).map(sharing => ({ ...sharing, access: "" }));
-
-            return {
-                meta: {
-                    allowPublicAccess: false,
-                    allowExternalAccess: false,
-                },
-                object: {
-                    id: "",
-                    displayName,
-                    externalAccess: false,
-                    publicAccess: "",
-                    userAccesses: buildSharings("user"),
-                    userGroupAccesses: buildSharings("userGroup"),
-                },
-            };
-        },
-        [settings]
-    );
-
-    const onUpdateSharingOptions = useCallback(
-        (setting: PermissionSetting) => {
-            return async ({ userAccesses: users, userGroupAccesses: userGroups }: ShareUpdate) => {
-                const buildPermission = (type: PermissionType, rule?: SharingRule[]) =>
-                    rule?.map(({ id, displayName }) => ({ id, displayName })) ??
-                    settings.getPermissions(setting, type);
-
-                const newSettings = settings
-                    .setPermissions(setting, "user", buildPermission("user", users))
-                    .setPermissions(setting, "userGroup", buildPermission("userGroup", userGroups));
-
-                await onChange(newSettings);
-            };
-        },
-        [onChange, settings]
-    );
-
     const buildSharingDescription = useCallback(
         (setting: PermissionSetting) => {
             const users = settings.getPermissions(setting, "user").length;
@@ -181,25 +130,13 @@ export default function SettingsFields(props: SettingsFieldsProps) {
                 />
             )}
 
-            {!!sharingDialogType && (
-                <ConfirmationDialog
-                    isOpen={true}
-                    fullWidth={true}
-                    onCancel={() => setSharingDialogType(null)}
-                    cancelText={i18n.t("Close")}
-                >
-                    <Sharing
-                        meta={buildMetaObject(sharingDialogType)}
-                        showOptions={{
-                            dataSharing: false,
-                            publicSharing: false,
-                            externalSharing: false,
-                            permissionPicker: false,
-                        }}
-                        onSearch={search}
-                        onChange={onUpdateSharingOptions(sharingDialogType)}
-                    />
-                </ConfirmationDialog>
+            {!!permissionsType && (
+                <PermissionsDialog
+                    onClose={() => setPermissionsType(null)}
+                    permissionsType={permissionsType}
+                    settings={settings}
+                    onChange={onChange}
+                />
             )}
 
             <h3 className={classes.title}>{i18n.t("Models")}</h3>
@@ -267,7 +204,7 @@ export default function SettingsFields(props: SettingsFieldsProps) {
             <h3 className={classes.title}>{i18n.t("Permissions")}</h3>
 
             <FormGroup className={classes.content} row={true}>
-                <ListItem button onClick={() => setSharingDialogType("generation")}>
+                <ListItem button onClick={() => setPermissionsType("generation")}>
                     <ListItemIcon>
                         <Icon>cloud_download</Icon>
                     </ListItemIcon>
@@ -276,7 +213,7 @@ export default function SettingsFields(props: SettingsFieldsProps) {
                         secondary={buildSharingDescription("generation")}
                     />
                 </ListItem>
-                <ListItem button onClick={() => setSharingDialogType("settings")}>
+                <ListItem button onClick={() => setPermissionsType("settings")}>
                     <ListItemIcon>
                         <Icon>settings</Icon>
                     </ListItemIcon>
@@ -298,11 +235,3 @@ const useStyles = makeStyles({
     duplicateTolerance: { margin: 0, marginRight: 15, width: 35 },
     duplicateToleranceLabel: { margin: 0, marginRight: 15, alignSelf: "center" },
 });
-
-function searchUsers(api: D2Api, query: string) {
-    const options = {
-        fields: { id: true, displayName: true },
-        filter: { displayName: { ilike: query } },
-    };
-    return api.metadata.get({ users: options, userGroups: options }).getData();
-}
